Refactor all models to game.models.*
